add wikipedia embeddings example
